Unify v1 and v2 test trees
c/parallel/ and c/parallel.v2/ shipped two near-identical copies of the
same test suite — both exercising the same C API, differing only in
backend plumbing. Any review fix had to be applied twice, and v1 quietly
drifted from v2 in some places (bug fixes ahead in v2, lint patterns
ahead in v1). Pick one source: c/parallel/test/.

c/parallel.v2/test/CMakeLists.txt now globs the test sources from
../../parallel/test/ and builds them with -DCCCL_C_PARALLEL_V2=1. The
v2 directory keeps only its CMakeLists.txt and freestanding/ (genuinely
v2-specific JIT-infra tests).

The shared sources gain three categories of compile-time switches:

  * #ifdef CCCL_C_PARALLEL_V2 in test_util.h to select between
    nvrtc->LTO-IR (v1) and hostjit->LLVM bitcode (v2) for compile(), and
    between CCCL_OP_LTOIR / CCCL_OP_LLVM_IR for make_operation(). v1's
    enum doesn't define CCCL_OP_LLVM_IR, so this can't be runtime-
    branched.

  * A new make_build_config() helper that uses value-init + explicit
    field assignments. v1's cccl_build_config has 4 fields, v2's has 6
    (enable_pch, verbose); the helper papers over the difference
    without #ifdef-ing each test site.

  * #ifndef CCCL_C_PARALLEL_V2 around blocks that don't apply to v2:
    test_segmented_reduce.cpp's guaranteed_max_segment_size variant
    (v2's API doesn't take that param, including in the base call),
    and test_unique_by_key.cpp's "fails to build for large types" test
    (v2 supports large keys). v2 picks up an equivalent positive
    large-keys test via #else.

v2's existing review fixes (prefix-only operator== in
three_way_partition, scratch-query pattern in binary_search, tag fix
plus keys/values verification in unique_by_key) are backported into the
unified files so they apply to v1 too.

In the same pass, removed dead SASS-check machinery:
inline constexpr bool check_ldl_stl_in_sass = false had been
unconditionally disabling the LDL/STL grep through inspect_sass since
nvbug 5243118. Drops the flag, build_traits<>, the if constexpr block
in algorithm_execution.h, every per-test should_check_sass() method,
and every SASS-only template parameter (with its non-default call
sites). inspect_sass() itself stays — it's the basis of the still-active
"Transform generates UBLKCP on SM90" assertion.

Net: 31 files changed, +232 / -9454.
