--- c/parallel.v2/test/test_identity.h
@@ -1,19 +1,0 @@
-//===----------------------------------------------------------------------===//
-//
-// Part of CUDA Experimental in CUDA C++ Core Libraries,
-// under the Apache License v2.0 with LLVM Exceptions.
-// See https://llvm.org/LICENSE.txt for license information.
-// SPDX-License-Identifier: Apache-2.0 WITH LLVM-exception
-// SPDX-FileCopyrightText: Copyright (c) 2025 NVIDIA CORPORATION & AFFILIATES.
-//
-//===----------------------------------------------------------------------===//
-
-#pragma once
-
-#ifdef TEST_IDENTITY_ENABLED
-template <typename T>
-__device__ T test_identity(T value)
-{
-  return value;
-}
-#endif
