+++ c/parallel/test/algorithm_execution.h
@@ -20,8 +20,6 @@
 #include <c2h/catch2_test_helper.h>
 #include <cccl/c/types.h>
 
-inline constexpr bool check_ldl_stl_in_sass = false;
-
 template <int device_id_ = 0>
 class BuildInformation
 {
@@ -81,24 +79,6 @@ public:
   }
 };
 
-template <typename Build, typename = void>
-struct build_traits
-{
-  static bool should_check_sass(int)
-  {
-    return true;
-  }
-};
-
-template <typename Build>
-struct build_traits<Build, std::void_t<decltype(Build::should_check_sass(0))>>
-{
-  static bool should_check_sass(int cc_major)
-  {
-    return Build::should_check_sass(cc_major);
-  }
-};
-
 template <typename BuildResultT,
           typename Build,
           typename Cleanup,
@@ -146,13 +126,6 @@ void AlgorithmExecute(std::optional<BuildCache>& cache, const std::optional<KeyT
       const auto& key_v = lookup_key.value();
       cache_v.insert(key_v, build);
     }
-  }
-
-  if constexpr (check_ldl_stl_in_sass && build_traits<Build>::should_check_sass(build_info.get_cc_major()))
-  {
-    const std::string sass = inspect_sass(build.cubin, build.cubin_size);
-    REQUIRE(sass.find("LDL") == std::string::npos);
-    REQUIRE(sass.find("STL") == std::string::npos);
   }
 
   CUstream null_stream = nullptr;
