+++ c/parallel/test/test_binary_search.cpp
@@ -77,19 +77,16 @@ struct binary_search_build
     return cccl_device_binary_search_build(
       build_ptr, mode, data, values, out, op, cc_major, cc_minor, cub_path, thrust_path, libcudacxx_path, ctk_path);
   }
-
-  static constexpr bool should_check_sass(int)
-  {
-    return false;
-  }
 };
 
 struct binary_search_run
 {
   template <typename... Ts>
-  CUresult operator()(BuildResultT build, void*, std::size_t*, cccl_binary_search_mode_t, Ts... args) const noexcept
+  CUresult operator()(
+    BuildResultT build, void* scratch, std::size_t* scratch_size, cccl_binary_search_mode_t, Ts... args) const noexcept
   {
-    return cccl_device_binary_search(build, args...);
+    *scratch_size = 1;
+    return (scratch) ? cccl_device_binary_search(build, args...) : CUDA_SUCCESS;
   }
 };
