+++ c/parallel/test/test_scan.cpp
@@ -63,7 +63,6 @@ auto& get_cache()
   return fixture<scan_build_cache_t, Tag>::get_or_create().get_value();
 }
 
-template <bool Disable75SassCheck = false, bool DisableForOtherArches = false>
 struct scan_build
 {
   CUresult operator()(
@@ -161,12 +160,6 @@ struct scan_build
       libcudacxx_path,
       ctk_path);
   }
-
-  static bool should_check_sass(int cc_major)
-  {
-    // TODO: add a check for NVRTC version; ref nvbug 5243118
-    return !(Disable75SassCheck && DisableForOtherArches) && (!Disable75SassCheck || cc_major > 7) && cc_major < 9;
-  }
 };
 
 struct scan_run
@@ -234,10 +227,7 @@ struct scan_run_no_init
   }
 };
 
-template <bool Disable75SassCheck    = false,
-          bool DisableForOtherArches = false,
-          typename BuildCache        = scan_build_cache_t,
-          typename KeyT              = std::string>
+template <typename BuildCache = scan_build_cache_t, typename KeyT = std::string>
 void scan(cccl_iterator_t input,
           cccl_iterator_t output,
           uint64_t num_items,
@@ -247,19 +237,11 @@ void scan(cccl_iterator_t input,
           std::optional<BuildCache>& cache,
           const std::optional<KeyT>& lookup_key)
 {
-  AlgorithmExecute<BuildResultT,
-                   scan_build<Disable75SassCheck, DisableForOtherArches>,
-                   scan_cleanup,
-                   scan_run,
-                   BuildCache,
-                   KeyT>(
+  AlgorithmExecute<BuildResultT, scan_build, scan_cleanup, scan_run, BuildCache, KeyT>(
     cache, lookup_key, inclusive, cccl_init_kind_t::CCCL_VALUE_INIT, input, output, num_items, op, init);
 }
 
-template <bool Disable75SassCheck    = false,
-          bool DisableForOtherArches = false,
-          typename BuildCache        = scan_build_cache_t,
-          typename KeyT              = std::string>
+template <typename BuildCache = scan_build_cache_t, typename KeyT = std::string>
 void scan(cccl_iterator_t input,
           cccl_iterator_t output,
           uint64_t num_items,
@@ -269,19 +251,11 @@ void scan(cccl_iterator_t input,
           std::optional<BuildCache>& cache,
           const std::optional<KeyT>& lookup_key)
 {
-  AlgorithmExecute<BuildResultT,
-                   scan_build<Disable75SassCheck, DisableForOtherArches>,
-                   scan_cleanup,
-                   scan_run_future_value,
-                   BuildCache,
-                   KeyT>(
+  AlgorithmExecute<BuildResultT, scan_build, scan_cleanup, scan_run_future_value, BuildCache, KeyT>(
     cache, lookup_key, inclusive, cccl_init_kind_t::CCCL_FUTURE_VALUE_INIT, input, output, num_items, op, init);
 }
 
-template <bool Disable75SassCheck    = false,
-          bool DisableForOtherArches = false,
-          typename BuildCache        = scan_build_cache_t,
-          typename KeyT              = std::string>
+template <typename BuildCache = scan_build_cache_t, typename KeyT = std::string>
 void scan(cccl_iterator_t input,
           cccl_iterator_t output,
           uint64_t num_items,
@@ -290,12 +264,7 @@ void scan(cccl_iterator_t input,
           std::optional<BuildCache>& cache,
           const std::optional<KeyT>& lookup_key)
 {
-  AlgorithmExecute<BuildResultT,
-                   scan_build<Disable75SassCheck, DisableForOtherArches>,
-                   scan_cleanup,
-                   scan_run_no_init,
-                   BuildCache,
-                   KeyT>(
+  AlgorithmExecute<BuildResultT, scan_build, scan_cleanup, scan_run_no_init, BuildCache, KeyT>(
     cache, lookup_key, inclusive, cccl_init_kind_t::CCCL_NO_INIT, input, output, num_items, op, nullptr);
 }
 
@@ -421,7 +390,7 @@ extern "C" __device__ void op(void* lhs_ptr, void* rhs_ptr, void* out_ptr) {
   auto& build_cache    = get_cache<Scan_CustomTypes_Fixture_Tag>();
   const auto& test_key = make_scan_key<pair>(false, cccl_init_kind_t::CCCL_VALUE_INIT);
 
-  scan<true>(input_ptr, output_ptr, num_items, op, init, false, build_cache, test_key);
+  scan(input_ptr, output_ptr, num_items, op, init, false, build_cache, test_key);
 
   std::vector<pair> expected(num_items, {0, 0});
   std::exclusive_scan(input.begin(), input.end(), expected.begin(), init.value, [](const pair& lhs, const pair& rhs) {
@@ -463,7 +432,7 @@ extern "C" __device__ void op(void* lhs_ptr, void* rhs_ptr, void* out_ptr) {
   auto& build_cache    = get_cache<Scan_CustomTypes_WellKnown_Fixture_Tag>();
   const auto& test_key = make_scan_key<pair>(false, cccl_init_kind_t::CCCL_VALUE_INIT);
 
-  scan<true>(input_ptr, output_ptr, num_items, op, init, false, build_cache, test_key);
+  scan(input_ptr, output_ptr, num_items, op, init, false, build_cache, test_key);
 
   std::vector<pair> expected(num_items, {0, 0});
   std::exclusive_scan(input.begin(), input.end(), expected.begin(), init.value, [](const pair& lhs, const pair& rhs) {
@@ -673,8 +642,7 @@ C2H_TEST("Scan works with floating point types", "[scan]", floating_point_types)
   auto& build_cache    = get_cache<Scan_FloatingPointTypes_Fixture_Tag>();
   const auto& test_key = make_scan_key<T>(false, cccl_init_kind_t::CCCL_VALUE_INIT);
 
-  // FIXME: figure out why scan spills to lmem for double
-  scan<std::is_same_v<T, double>, true>(input_ptr, output_ptr, num_items, op, init, false, build_cache, test_key);
+  scan(input_ptr, output_ptr, num_items, op, init, false, build_cache, test_key);
 
   const std::vector<T> output = output_ptr;
   std::vector<T> expected(num_items);
@@ -709,13 +677,9 @@ C2H_TEST("Scan works with C++ source operations using custom headers", "[scan]")
   value_t<T> init{T{42}};
 
   // Test _ex version with custom build configuration
-  cccl_build_config config;
-  const char* extra_flags[]      = {"-DTEST_IDENTITY_ENABLED"};
-  const char* extra_dirs[]       = {TEST_INCLUDE_PATH};
-  config.extra_compile_flags     = extra_flags;
-  config.num_extra_compile_flags = 1;
-  config.extra_include_dirs      = extra_dirs;
-  config.num_extra_include_dirs  = 1;
+  const char* extra_flags[] = {"-DTEST_IDENTITY_ENABLED"};
+  const char* extra_dirs[]  = {TEST_INCLUDE_PATH};
+  cccl_build_config config  = make_build_config(extra_flags, 1, extra_dirs, 1);
 
   // Build with _ex version
   cccl_device_scan_build_result_t build;
