+++ c/parallel/test/test_segmented_sort.cpp
@@ -73,14 +73,8 @@ auto& get_cache()
   return fixture<segmented_sort_build_cache_t, Tag>::get_or_create().get_value();
 }
 
-template <bool DisableSassCheckOnSm120 = false>
 struct segmented_sort_build
 {
-  static bool should_check_sass(int cc_major)
-  {
-    return !(DisableSassCheckOnSm120 && cc_major >= 12);
-  }
-
   CUresult operator()(
     BuildResultT* build_ptr,
     cccl_sort_order_t sort_order,
@@ -151,9 +145,7 @@ struct segmented_sort_run
   }
 };
 
-template <bool DisableSassCheckOnSm120 = false,
-          typename BuildCache          = segmented_sort_build_cache_t,
-          typename KeyT                = std::string>
+template <typename BuildCache = segmented_sort_build_cache_t, typename KeyT = std::string>
 void segmented_sort(
   cccl_sort_order_t sort_order,
   cccl_iterator_t keys_in,
@@ -169,12 +161,7 @@ void segmented_sort(
   std::optional<BuildCache>& cache,
   const std::optional<KeyT>& lookup_key)
 {
-  AlgorithmExecute<BuildResultT,
-                   segmented_sort_build<DisableSassCheckOnSm120>,
-                   segmented_sort_cleanup,
-                   segmented_sort_run,
-                   BuildCache,
-                   KeyT>(
+  AlgorithmExecute<BuildResultT, segmented_sort_build, segmented_sort_cleanup, segmented_sort_run, BuildCache, KeyT>(
     cache,
     lookup_key,
     sort_order,
@@ -200,11 +187,10 @@ C2H_TEST("segmented_sort can sort keys-only", "[segmented_sort][keys_only]", tes
   using T     = c2h::get<0, TestType>;
   using key_t = typename T::KeyT;
 
-  constexpr auto this_test_params            = T();
-  constexpr bool is_descending               = this_test_params.is_descending();
-  constexpr auto order                       = is_descending ? CCCL_DESCENDING : CCCL_ASCENDING;
-  constexpr bool is_overwrite_okay           = this_test_params.is_overwrite_okay();
-  constexpr bool disable_sass_check_on_sm120 = std::is_same_v<key_t, c2h::get<3, key_types>>;
+  constexpr auto this_test_params  = T();
+  constexpr bool is_descending     = this_test_params.is_descending();
+  constexpr auto order             = is_descending ? CCCL_DESCENDING : CCCL_ASCENDING;
+  constexpr bool is_overwrite_okay = this_test_params.is_overwrite_okay();
 
   const std::size_t n_segments   = GENERATE(0, 13, take(2, random(1 << 10, 1 << 12)));
   const std::size_t segment_size = GENERATE(1, 12, take(2, random(1 << 10, 1 << 12)));
@@ -287,7 +273,7 @@ C2H_TEST("segmented_sort can sort keys-only", "[segmented_sort][keys_only]", tes
 
   int selector = -1;
 
-  segmented_sort<disable_sass_check_on_sm120>(
+  segmented_sort(
     order,
     keys_in_ptr,
     keys_out_ptr,
@@ -330,11 +316,10 @@ C2H_TEST("segmented_sort can sort key-value pairs", "[segmented_sort][key_value]
   using T     = c2h::get<0, TestType>;
   using key_t = typename T::KeyT;
 
-  constexpr auto this_test_params            = T();
-  constexpr bool is_descending               = this_test_params.is_descending();
-  constexpr auto order                       = is_descending ? CCCL_DESCENDING : CCCL_ASCENDING;
-  constexpr bool is_overwrite_okay           = this_test_params.is_overwrite_okay();
-  constexpr bool disable_sass_check_on_sm120 = !std::is_same_v<key_t, c2h::get<0, key_types>>;
+  constexpr auto this_test_params  = T();
+  constexpr bool is_descending     = this_test_params.is_descending();
+  constexpr auto order             = is_descending ? CCCL_DESCENDING : CCCL_ASCENDING;
+  constexpr bool is_overwrite_okay = this_test_params.is_overwrite_okay();
 
   const std::size_t n_segments   = GENERATE(0, 13, take(2, random(1 << 10, 1 << 12)));
   const std::size_t segment_size = GENERATE(1, 12, take(2, random(1 << 10, 1 << 12)));
@@ -387,7 +372,7 @@ C2H_TEST("segmented_sort can sort key-value pairs", "[segmented_sort][key_value]
 
   int selector = -1;
 
-  segmented_sort<disable_sass_check_on_sm120>(
+  segmented_sort(
     order,
     keys_in_ptr,
     keys_out_ptr,
@@ -599,10 +584,7 @@ C2H_TEST("SegmentedSort works with variable segment sizes", "[segmented_sort][va
   constexpr bool is_descending     = this_test_params.is_descending();
   constexpr auto order             = is_descending ? CCCL_DESCENDING : CCCL_ASCENDING;
   constexpr bool is_overwrite_okay = this_test_params.is_overwrite_okay();
-  constexpr bool disable_sass_check_on_sm120 =
-    std::is_same_v<key_t, c2h::get<1, key_types>> || std::is_same_v<key_t, c2h::get<2, key_types>>;
-
-  const std::size_t n_segments = GENERATE(20, 600);
+  const std::size_t n_segments     = GENERATE(20, 600);
 
   // Create variable segment sizes
   const std::vector<std::size_t> base_pattern = {
@@ -662,7 +644,7 @@ C2H_TEST("SegmentedSort works with variable segment sizes", "[segmented_sort][va
 
   int selector = -1;
 
-  segmented_sort<disable_sass_check_on_sm120>(
+  segmented_sort(
     order,
     keys_in_ptr,
     keys_out_ptr,
